requirements.txt
pyodbc
python-dotenv
langchain
langchain-community
langchain_groq
streamlit
pandas 
fuzzywuzzy
sqlalchemy